[FEATURE] Add extended YamlPersistenceManager
This YamlPersistenceManager first loads forms from Packages/*/*/Data/Forms/.
The first form found in a package will be used for rendering. If no form is
found the global TYPO3.Form.yamlPersistenceManager.savePath setting will
be used.

New forms are always created in the savePath, and changes to a form in a
package will also be saved in the global path to prevent changes to the
form in the original package.